+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Converter for Lithuania (EuroCrops)
 - Converter for Switzerland
 
 ### Fixed